# Faster version for List.last/1