@AndrewDryga 2017-09-01
Dump. Had a typo in my bench, it's much slower for lists with more than 300000 elements.