@AndrewDryga 2017-09-01
@josevalim what is your opinion on this kind of optimizations? For eg.:
```
Settings:
  duration:      3.0 s

## ListBench
[20:01:06] 1/12: Elixir 10
[20:01:15] 2/12: Elixir 10000
[20:01:21] 3/12: Elixir 100000
[20:01:27] 4/12: Elixir 1000000
[20:01:33] 5/12: Opt 10
[20:01:37] 6/12: Opt 10000
[20:01:41] 7/12: Opt 100000
[20:01:44] 8/12: Opt 1000000
[20:01:48] 9/12: Reverse 10
[20:01:55] 10/12: Reverse 10000
[20:01:58] 11/12: Reverse 100000
[20:02:02] 12/12: Reverse 1000000

Finished in 66.62 seconds

## ListBench
benchmark name   iterations   average time 
Opt 10            100000000   0.04 µs/op
Reverse 10        100000000   0.06 µs/op
Elixir 10         100000000   0.08 µs/op
Reverse 10000        100000   30.08 µs/op
Opt 10000            100000   30.73 µs/op
Elixir 10000         100000   54.20 µs/op
Opt 100000            10000   317.39 µs/op
Reverse 100000        10000   330.47 µs/op
Elixir 100000         10000   503.03 µs/op
Opt 1000000            1000   3689.87 µs/op
Elixir 1000000         1000   5390.93 µs/op
Reverse 1000000         500   16068.88 µs/op
```

Code:
```
defmodule ListBench do
  use Benchfella

  bench "Elixir 10", [data: gen_list(10)] do
    List.last(data)
    :ok
  end

  bench "Elixir 10000", [data: gen_list(10000)] do
    List.last(data)
    :ok
  end

  bench "Elixir 100000", [data: gen_list(100000)] do
    List.last(data)
    :ok
  end

  bench "Elixir 1000000", [data: gen_list(1000000)] do
    List.last(data)
    :ok
  end

  bench "Reverse 10", [data: gen_list(10)] do
    last_rev(data)
    :ok
  end

  bench "Reverse 10000", [data: gen_list(10000)] do
    last_rev(data)
    :ok
  end

  bench "Reverse 100000", [data: gen_list(100000)] do
    last_rev(data)
    :ok
  end

  bench "Reverse 1000000", [data: gen_list(1000000)] do
    last_rev(data)
    :ok
  end

  bench "Opt 10", [data: gen_list(10)] do
    last_opt(data)
    :ok
  end

  bench "Opt 10000", [data: gen_list(10000)] do
    last_opt(data)
    :ok
  end

  bench "Opt 100000", [data: gen_list(100000)] do
    last_opt(data)
    :ok
  end

  bench "Opt 1000000", [data: gen_list(1000000)] do
    last_opt(data)
    :ok
  end

  defp gen_list(n) do
    for i <- 0..n, do: i
  end

  defp last_rev(list), do: :erlang.hd(:lists.reverse(list))

  def last_opt([]), do: nil
  def last_opt([head]), do: head
  def last_opt([_, _, _, _, _, _, _, _, _, head]), do: head # Less function calls for longer lists, could be replaced with code generation
  def last_opt([_, _, _, _, _, _, _, _, _, _ | tail]), do: last_opt(tail)
end
```

I can dig deeper in each List function and check is it possible to impove them, just for fun.