@michalmuskala 2017-09-01
Fiddling with the `:inline` setting of the `@compile` attribute could produce similar results.