@AndrewDryga 2017-09-01
@michalmuskala `List` module already inlines calls to `:lists`..