@michalmuskala 2017-09-01
Inlining the `:lists` module and inlining in general are two separate things. Erlang docs have a bit more on [http://erlang.org/doc/man/compile.html#id62277](this), but it's still not much.

In case of recursive functions inlining them effectively means unrolling the loop.